+++ source/commands/utils/sharedDangerfileArgs.ts
@@ -27,7 +27,7 @@ export interface SharedCLI extends program.CommanderStatic {
   failOnErrors: boolean
   /** Use GitHub Checks */
   useGithubChecks: boolean
-  /** Ignore invalid inline-comments, for instance a comment with a line that was not changed */
+  /** Ignore inline-comments that are in lines which were not changed */
   noOutOfDiffComments: boolean
 }
 
@@ -45,8 +45,4 @@ export default (command: any) =>
     .option("-u, --passURLForDSL", "[dev] Use a custom URL to send the Danger DSL into the sub-process")
     .option("-f, --failOnErrors", "Fail if there are fails in the danger report", false)
     .option("--use-github-checks", "Use GitHub Checks", false)
-    .option(
-      "--noOutOfDiffComments",
-      "Ignore invalid inline-comments, for instance a comment with a line that was not changed",
-      false
-    )
+    .option("--noOutOfDiffComments", "Ignore inline-comments that are in lines which were not changed", false)